requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
streamlit
python-box
lida[transformers]
python-dotenv
streamlit